refactor: reuse rig test support helpers
